+++ common/src/main/java/io/github/mortuusars/sootychimneys/SootyChimneys.java
@@ -6,7 +6,6 @@ import io.github.mortuusars.sootychimneys.data.Chimney;
 import io.github.mortuusars.sootychimneys.data.ChimneyTypes;
 import io.github.mortuusars.sootychimneys.recipe.SootScrapingRecipe;
 import net.minecraft.core.Registry;
-import net.minecraft.core.particles.SimpleParticleType;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
@@ -246,7 +245,9 @@ public final class SootyChimneys {
     }
 
     public static class Stats {
-        private static final Map<ResourceLocation, StatFormatter> STATS = new HashMap<>();
+        public static final Map<ResourceLocation, StatFormatter> STATS = new HashMap<>();
+
+        public static final ResourceLocation SOOT_SCRAPED = register(resource("soot_scraped"), StatFormatter.DEFAULT);
 
         @SuppressWarnings("SameParameterValue")
         private static ResourceLocation register(ResourceLocation location, StatFormatter formatter) {
@@ -276,11 +277,15 @@ public final class SootyChimneys {
 
     public static class Tags {
         public static class Items {
-            public static final TagKey<Item> CHIMNEYS = TagKey.create(Registries.ITEM, ResourceLocation.parse("c:chimneys"));
+            public static final TagKey<Item> CHIMNEYS = TagKey.create(Registries.ITEM, resource("chimneys"));
+            public static final TagKey<Item> SOOTY_CHIMNEYS = TagKey.create(Registries.ITEM, resource("sooty_chimneys"));
+            public static final TagKey<Item> C_CHIMNEYS = TagKey.create(Registries.ITEM, ResourceLocation.parse("c:chimneys"));
         }
 
         public static class Blocks {
-            public static final TagKey<Block> CHIMNEYS = TagKey.create(Registries.BLOCK, ResourceLocation.parse("c:chimneys"));
+            public static final TagKey<Block> CHIMNEYS = TagKey.create(Registries.BLOCK, resource("chimneys"));
+            public static final TagKey<Block> SOOTY_CHIMNEYS = TagKey.create(Registries.BLOCK, resource("sooty_chimneys"));
+            public static final TagKey<Block> C_CHIMNEYS = TagKey.create(Registries.BLOCK, ResourceLocation.parse("c:chimneys"));
             public static final TagKey<Block> SMOKE_BLOCKING = TagKey.create(Registries.BLOCK, SootyChimneys.resource("smoke_blocking"));
             public static final TagKey<Block> SMOKE_BOOSTING = TagKey.create(Registries.BLOCK, SootyChimneys.resource("smoke_boosting"));
         }