+++ common/src/main/java/io/github/mortuusars/sootychimneys/data/ChimneyType.java
@@ -1,6 +1,5 @@
 package io.github.mortuusars.sootychimneys.data;
 
-import io.github.mortuusars.sootychimneys.core.ChimneyShape;
-import io.github.mortuusars.sootychimneys.core.smoke.SmokeProperties;
+import io.github.mortuusars.sootychimneys.data.smoke.SmokeProperties;
 
 public record ChimneyType(SmokeProperties smokeProperties, ChimneyShape shape) { }