+++ common/src/main/java/io/github/mortuusars/sootychimneys/data/smoke/ParticleOrigin.java
@@ -1,4 +1,4 @@
-package io.github.mortuusars.sootychimneys.core.smoke;
+package io.github.mortuusars.sootychimneys.data.smoke;
 
 import org.joml.Vector3f;