+++ neoforge/src/main/java/io/github/mortuusars/sootychimneys/neoforge/SootyChimneysNeoForge.java
@@ -1,6 +1,10 @@
 package io.github.mortuusars.sootychimneys.neoforge;
 
 import io.github.mortuusars.sootychimneys.Config;
+import net.minecraft.core.Registry;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.stats.StatType;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
@@ -14,6 +18,11 @@ import net.neoforged.fml.loading.FMLEnvironment;
 public final class SootyChimneysNeoForge {
     public SootyChimneysNeoForge(ModContainer container) {
         SootyChimneys.init();
+
+        SootyChimneys.Stats.STATS.forEach((location, formatter) -> {
+            RegisterImpl.CUSTOM_STATS.register(location.getPath(), () -> location);
+        });
+
         container.registerConfig(ModConfig.Type.COMMON, Config.Common.SPEC);
         container.registerConfig(ModConfig.Type.CLIENT, Config.Client.SPEC);
 
@@ -33,6 +42,7 @@ public final class SootyChimneysNeoForge {
         RegisterImpl.WORLD_GEN_FEATURES.register(modEventBus);
         RegisterImpl.DATA_COMPONENT_TYPES.register(modEventBus);
         RegisterImpl.PARTICLE_TYPES.register(modEventBus);
+        RegisterImpl.CUSTOM_STATS.register(modEventBus);
 
         if (FMLEnvironment.dist == Dist.CLIENT) {
             SootyChimneysNeoForgeClient.init(container);